fixed kNN for kNNBasic on issue #131
